requirements.txt
certifi==2019.11.28
chardet==3.0.4
dnspython==1.16.0
get==2019.4.13
idna==2.9
mysql-connector-python==8.0.19
post==2019.4.13
protobuf==3.6.1
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
six==1.14.0
urllib3==1.25.8
